Ultimos cambios

+++ database/migrations/2025_10_18_062307_create_calendarios_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
                 $table->time('HoraFin')->nullable();
                 $table->string('Ubicacion', 200)->nullable();
                 $table->unsignedBigInteger('OrganizadorID')->nullable();
-                $table->unsignedBigInteger('ProyectoID')->nullable();
+                $table->unsignedInteger('ProyectoID')->nullable();
                 
                 $table->foreign('OrganizadorID')->references('MiembroID')->on('miembros')->onDelete('set null');
                 $table->foreign('ProyectoID')->references('ProyectoID')->on('proyectos')->onDelete('set null');
